requirements.txt
djangorestframework
drf-yasg
psycopg2-binary
python-decouple
